Extract pool to its own module.
