+++ src/rust/engine/fs/src/snapshot.rs
@@ -1,3 +1,6 @@
+// Copyright 2017 Pants project contributors (see CONTRIBUTORS.md).
+// Licensed under the Apache License, Version 2.0 (see LICENSE).
+
 use PathStat;
 use hash::Fingerprint;
 use std::fmt;
